fix(echidna): fix URL param loading — pre-load in main.tsx + resize kick
Move URL param (?load=) and postMessage loading to main.tsx entry point,
loading data into the Zustand store BEFORE React mounts. This ensures the
R3F Canvas sees populated data on first render.

Also dispatch a resize event 100ms after mount to kick-start the R3F
render loop, which otherwise defers its first frame until user interaction.

Add InitialInvalidator component + CameraFitter for auto-framing.
